src/server/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
httpx